Implement Saving and restoring of mobile numbers
I have finished the saving of the entered number's to the memory
and retrival of it from memory.
I changed the key of the shared preference used for this to a more
suitable one
I also made the color change of the main button which
asks the user to set his emergency numbers to a more pleasing Blue.
The label of the emergency number activity is more Informative and
better looking, by naming it as Emergency Numbers

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>