Fixed storing to db problem
we werent able to connect to a external database (vishnu's localhost)
from the python script in my computer. even though vishnu's localhost
was forwarded using ngrok.

So he made a php page which takes the values to be added in a get
request and stores them into the db.

so to store data I just call the url with the date and revolutions as
parameters.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>