Project Done
This project is over and this is how the app was at the end of it.
Sadly towards the end even as I added more features to the app, some of
them stopped working. Even now I do not know why the app no longer reads
the messages out to me (SMS's).

I will be writing a blog post (hopefully soon) detailing the project.
I will then be cleaning this repo up. Until then

Janna

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>